+++ src/modules/visualBuilder/dataUpdate.js
@@ -1,4 +1,4 @@
-export const updateFormData = (element, formStyle) => {
+export const updateFormData = (element, formStyle, payload, isPreview = false) => {
   // Update the element style
   if (formStyle.style !== undefined) {
     Object.keys(formStyle.style).forEach((property) => {
@@ -18,15 +18,25 @@ export const updateFormData = (element, formStyle) => {
 
   // Update element text
   if (formStyle.text !== undefined) {
-    element.innerText = formStyle.text
+    element.innerText = isPreview ? formStyle.text.text : formStyle.text
   }
 
   // Handle element onClick
   if (formStyle.clickDetails !== undefined) {
     const url = formStyle.clickDetails.clickUrl
     element.onclick = formStyle.clickDetails.newTab
-      ? () => window.open(url, '_blank').focus()
-      : () => { window.location.href = url }
+      ? () => {
+        if (!isPreview) {
+          window.clevertap.raiseNotificationClicked(payload)
+        }
+        window.open(url, '_blank').focus()
+      }
+      : () => {
+        if (!isPreview) {
+          window.clevertap.raiseNotificationClicked(payload)
+        }
+        window.location.href = url
+      }
   }
 
   // Set the image source